split setup helpers
