+++ examples/setup/README.md
@@ -9,6 +9,10 @@ Client secrets are generated by Magistrala during client creation and copied
 into the run state so later messaging examples can authenticate as those
 clients.
 
+The flat files in `examples/*.ts` remain short API references with placeholder
+values. The files in `examples/setup/*.ts` are runnable scenarios for a live
+Magistrala instance.
+
 ## E2E Runs
 
 ```bash
@@ -30,6 +34,7 @@ dependencies first:
 npx ts-node examples/setup/clients.ts
 npx ts-node examples/setup/channels.ts
 npx ts-node examples/setup/groups.ts
+npx ts-node examples/setup/rules.ts
 npx ts-node examples/setup/reports.ts
 ```
 
@@ -58,6 +63,14 @@ unless you set:
 
 ```bash
 MG_RUN_DESTRUCTIVE=true
+```
+
+Terminal output redacts passwords, tokens, and secrets by default. The raw
+values are still saved in the gitignored run state file. To print raw secrets in
+the terminal, run with:
+
+```bash
+MG_SETUP_SHOW_SECRETS=true
 ```
 
 By default the setup resumes from the local `.runs/<run-id>.json` file and does
