+++ examples/setup/helpers/reports.ts
@@ -1,0 +1,145 @@
+// Copyright (c) Abstract Machines
+// SPDX-License-Identifier: Apache-2.0
+
+import type {
+  Aggregation,
+  Channel,
+  Format,
+  PAT,
+  ReportConfig,
+} from "../../../src/sdk";
+import type { DomainResources, StoredClient } from "../state";
+import { saveState } from "../state";
+import {
+  type ScenarioContext,
+  isTransportFailure,
+  prefixFor,
+  requireDomainId,
+  requireId,
+  requireToken,
+  runStep,
+  sampleMessageNames,
+  shouldAdoptExisting,
+  sleep,
+  tryStep,
+} from "./common";
+
+const findReportConfigByName = async (
+  context: ScenarioContext,
+  domainId: string,
+  name: string
+): Promise<ReportConfig | undefined> => {
+  const { sdk, state, logger } = context;
+  const token = requireToken(state).access_token;
+  const configs = await tryStep(logger, `reportConfig.${name}.lookup`, () => sdk.Reports.listConfigs(
+    domainId,
+    {
+      offset: 0,
+      limit: 100,
+    },
+    token
+  ));
+  return (configs?.report_configs ?? []).find((config) => config.name === name);
+};
+
+export const createPat = async (
+  context: ScenarioContext,
+  index: number
+): Promise<PAT> => {
+  const { config, sdk, state, logger } = context;
+  const existing = state.pats[index];
+  if (existing) {
+    logger.skipped(`pat.${index}`, "already present in run state");
+    return existing;
+  }
+
+  const token = requireToken(state).access_token;
+  const pat = await runStep(logger, `pat.${index}.create`, () => sdk.PATs.create(
+    `${prefixFor(config, state)}-pat-${index + 1}`,
+    "24h",
+    token,
+    "Created by examples/setup"
+  ));
+  state.pats.push(pat);
+  saveState(config, state);
+  return pat;
+};
+
+export const createReportConfig = async (
+  context: ScenarioContext,
+  domain: DomainResources,
+  client: StoredClient,
+  channel: Channel,
+  index: number
+): Promise<ReportConfig> => {
+  const { config, sdk, state, logger } = context;
+  const existing = domain.reportConfigs[index];
+  if (existing) {
+    logger.skipped(`reportConfig.${index}`, "already present in run state");
+    return existing;
+  }
+
+  const token = requireToken(state).access_token;
+  const domainId = requireDomainId(domain);
+  const name = `${prefixFor(config, state)}-report-${index + 1}`;
+  if (shouldAdoptExisting()) {
+    const existingConfig = await findReportConfigByName(context, domainId, name);
+    if (existingConfig) {
+      logger.resource(`reportConfig.${index}.adopt`, existingConfig);
+      domain.reportConfigs.push(existingConfig);
+      saveState(config, state);
+      return existingConfig;
+    }
+  }
+
+  const reportConfig: ReportConfig = {
+    name,
+    description: "Created by examples/setup",
+    config: {
+      title: `Setup Report ${index + 1}`,
+      from: "now()-1h",
+      to: "now()",
+      aggregation: {
+        agg_type: "none" as Aggregation,
+        interval: "1m",
+      },
+      file_format: "csv" as Format,
+      timezone: "UTC",
+    },
+    metrics: [
+      {
+        name: sampleMessageNames[index % sampleMessageNames.length],
+        channel_id: requireId(channel, "channel"),
+        client_ids: [requireId(client.data, "client")],
+        subtopic: "",
+        protocol: "http",
+        format: "",
+      },
+    ],
+    email: {
+      to: [state.user?.credentials.email ?? "user@example.com"],
+      subject: `Setup report ${index + 1}`,
+      content: "Report generated by examples/setup",
+    },
+  };
+
+  try {
+    const created = await runStep(logger, `reportConfig.${index}.add`, () => sdk.Reports.addConfig(domainId, reportConfig, token));
+    domain.reportConfigs.push(created);
+    saveState(config, state);
+    return created;
+  } catch (error) {
+    logger.error(`reportConfig.${index}.add`, error);
+    if (isTransportFailure(error)) {
+      await sleep(1000);
+    }
+    const recovered = await findReportConfigByName(context, domainId, name);
+    if (!recovered) {
+      throw error;
+    }
+    logger.resource(`reportConfig.${index}.recover`, recovered);
+    domain.reportConfigs.push(recovered);
+    saveState(config, state);
+    return recovered;
+  }
+};
