+++ examples/setup/rules.ts
